Add GetSummaryInformation() to VMM and VM APIs
The latter returns for all VMS, the former for the version associated with the VM.
Also, improve support for interval datetimes and introduce Duration support.

Unfortunately, time. Time does not provide access to raw seconds and nanoseconds,
but it does provide them relative to Unix epoch. Intervals are therefore encoded
as a delta from Unix time when stored in a time.Time. Callers can subtract Unix
epoc (time.Unix(0,0)), but the result is capped at 292 years due to the Duration
being stored as a total nanos value. Alternatively, they can use .Unix() and
.Nanosecond() to access the second and partial second as nanos value, providing
a complete view of all possible DateTime interval values (Max is 273972 years).

Signed-off-by: Jason T. Greene <jason.greene@redhat.com>