pkg/kvp: use uint32 for hvKvpMsgRet.error, fixing 32-bit builds
Error codes are represented as 32-bit hex constants with the high bit set,
and these can't be assigned to a signed 32-bit field.  The error field is
always 32 bits in the ABI, but we were making it 64 bits on 64-bit systems.
Since we're on little-endian machines, this didn't change the outcome
(except to make the struct too large, which is apparently tolerated by the
kernel ABI), but it did cause a compilation failure on 32-bit builds.

Fix by changing the field to a uint32.

Signed-off-by: Benjamin Gilbert <bgilbert@redhat.com>