Use PowerShell command to interact with Hyper-V
Replace wmi calls with powershell commands. This should improve error messages and remove need of executing this code with escalating privileges. The workflow should be the same, the most changes is in 'MemorySettings' memory field names and types, it should be set in bytes. And 'ProcessorSettings' has other name for processors count field.

 Signed-off-by: Yevhen Vydolob <yvydolob@redhat.com>

Signed-off-by: Yevhen Vydolob <yvydolob@redhat.com>